kvserver/reports: minor code move
Some testing functions were in a random test file that wasn't even
calling them. They're high-profile functions, so I've created a
dedicated helpers_test.go for them.

Release note: None
